Enable xrdp for redaht

+++ Artifacts/linux-rdp/install-xrdp.sh
@@ -1,58 +1,21 @@
 #!/bin/sh
 
-xfce()
-{
-  apt-get install xrdp -y -qq
-  apt-get install xfce4 -y -qq
-(
-cat <<'EOF'
-#!/bin/sh
+case $(uname -m) in
+"x86_64" | "amd64")
+  ;;
+*)
+  echo "This only works on amd64" 1>&2
+  exit -1
+  ;;
+esac
 
-if [ -r /etc/default/locale ]; then
-  . /etc/default/locale
-  export LANG LANGUAGE
-fi
-
-startxfce4
-EOF
-) > /etc/xrdp/startwm.sh
-}
-
-unity()
-{
-  if [ $(lsb_release -si) != "Ubuntu" ]; then
-    (>&2 echo "This GUI shell only works on ubuntu")
-    exit -1
-  fi
-  apt-get install ubuntu-desktop -y -qq
-  if [ $(uname -m) = "x86_64" ]; then
-    ARCH=amd64
-  else
-    ARCH=i386
-  fi
-  TIGERVNCPATH=https://dl.bintray.com/tigervnc/stable/ubuntu-$(lsb_release -sr)LTS/$ARCH/tigervncserver_1.7.0-1ubuntu1_$ARCH.deb
-  curl -L $TIGERVNCPATH -o tigervncserver.deb
-  dpkg -i tigervncserver.deb
-  apt-get install -fy -qq
-  apt-get install xrdp -y -qq
-(
-cat <<'EOF'
-#!/bin/sh
-
-if [ -r /etc/default/locale ]; then
-  . /etc/default/locale
-  export LANG LANGUAGE
-fi
-
-unity
-EOF
-) > /etc/xrdp/startwm.sh
-}
-
-mate()
-{
-  apt-get install xrdp -y -qq
-  apt-get install mate-core mate-desktop-environment mate-notification-daemon mate-tweak mate-dock-applet -y -qq
+if [ -f /etc/os-release ]; then
+  . /etc/os-release
+  case "$ID:$VERSION_ID" in
+  "ubuntu:14.04" | "ubuntu:16.04" | "ubuntu:16.10")
+    apt-get update
+    apt-get install xrdp -y -qq
+    apt-get install mate-core mate-desktop-environment mate-notification-daemon mate-tweak mate-dock-applet -y -qq
 (
 cat <<'EOF'
 #!/bin/sh
@@ -65,9 +28,29 @@ fi
 mate-session
 EOF
 ) > /etc/xrdp/startwm.sh
-}
-
-apt-get update
-eval $1
-service xrdp restart
-shutdown -r +1
+    shutdown -r +1
+    ;;
+  "rhel:7.0" | "centos:7")
+    wget https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm
+    wget http://li.nux.ro/download/nux/dextop/el7/x86_64/nux-dextop-release-0-5.el7.nux.noarch.rpm
+    yum update
+    yum -y install epel-release-latest-7.noarch.rpm
+    yum -y install nux-dextop-release-0-5.el7.nux.noarch.rpm
+    yum -y install mate-desktop
+    yum -y install xrdp tigervnc-server
+    systemctl enable xrdp.service
+    firewall-cmd --permanent --zone=public --add-port=3389/tcp
+    firewall-cmd --reload
+    chcon --type=bin_t /usr/sbin/xrdp
+    chcon --type=bin_t /usr/sbin/xrdp-sesman
+    shutdown -r +1
+    ;;
+  *)
+    echo "Unknown or not supported distro" 1>&2
+    exit -1
+    ;;
+  esac
+else
+  echo "Could not discover distro" 1>&2
+  exit -1
+fi
